third_party/blink/web_tests/platform/linux/virtual/webnn-service-with-gpu/external/wpt/webnn/conformance_tests/add.https.any_gpu-expected.txt

This is a testharness.js-based test.
[FAIL] [required] add float16 large inputs
  assert_less_than_equal: assert_array_approx_equals_ulp: test add float16 actual 0 should be close enough to expected 166.5 by ULP distance: expected a number less than or equal to 1 but got 22836
Harness: the test ran to completion.
